+++ src/api/articles/index.ts
@@ -21,6 +21,7 @@ const articlesQuerySchema = z.object({
 });
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+  console.log("req: ", req);
   if (req.method !== "GET") {
     return res.status(405).json({ message: "Method Not Allowed" });
   }
