+++ src/pages/api/articles.page.ts
@@ -1,0 +1,3 @@
+import handler from "@/api/articles";
+
+export default handler;
