+++ src/pages/api/preview.page.ts
@@ -1,0 +1,3 @@
+import handler from "@/api/preview";
+
+export default handler;
